requirements.txt
# 核心依赖
nonebot2>=2.4.2,<3.0.0
nonebot-adapter-onebot>=2.4.6,<3.0.0
nonebot-plugin-apscheduler>=0.5.0,<0.6.0
nonebot-plugin-uninfo>=0.8.0,<0.9.0
nonebot-plugin-alconna>=0.58.2,<0.59.0

# FastAPI 等相关依赖
fastapi>=0.100.0,<1.0.0
uvicorn>=0.23.0,<1.0.0
httpx>=0.24.1,<1.0.0
websockets>=11.0.3,<16.0.0
openai>=1.0.0,<2.0.0
python-dotenv>=0.21.0,<2.0.0
pyyaml>=6.0.1,<7.0.0
pandas>=2.0.3,<3.0.0
openpyxl>=3.1.2,<4.0.0

# 系统依赖
psutil>=5.9.0,<8.0.0

# WebUI 依赖
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.5,<1.0.0
requests>=2.28.1,<3.0.0
pydantic>=2.0.0,<3.0.0,!=2.5.0,!=2.5.1,!=2.10.0,!=2.10.1
pydantic-settings>=2.0.0,<3.0.0
passlib>=1.7.4,<2.0.0
bcrypt>=4.0.1,<5.0.0
PyJWT>=2.0.0,<3.0.0
tortoise-orm>=0.17.0,<1.0.0
pytz>=2022.1
tzlocal>=4.2,<6.0.0

# 词云依赖
jieba>=0.42.1,<1.0.0

# 其他缺失依赖
neomodel>=5.0.0,<6.0.0
